+++ jest.config.mjs
@@ -1,0 +1,7 @@
+export default {
+  testMatch: [
+    "**/__tests__/**/*.test.?(m)[jt]s?(x)"
+  ],
+  verbose: true,
+  testTimeout: 120000,
+}