+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Added `"type": "module"` to `package.json`
 - Updated Rollup and TypeScript.
+- Excluded sharp from the bundle.
+
+    rollup.config.js:
+    ```js
+    external: [
+        ...builtins,
+        'sharp',
+    ],
+    ```
 - Renamed `File` class to `FileUtils`.  
     Renamed `Media` class to `MediaUtils`.  
     Added new utility classes: `ProcessUtils` and `ImageUtils`.
@@ -36,9 +45,6 @@ All notable changes to this project will be documented in this file.
     | `MediaUtils` | Renamed and changed parameters. Added options for file owner and group for files. | `writeDataUrlToFile(outputPath: string, dataUrl: string, permission: number = 0o755): string` | `writeImage(outputPath: string, dataUrl: string, options: WriteImageOptions = {mode: 0o755})` |
     | `ProcessUtils` | New method added. | - | `getUid(username: string): number` |
     | `ProcessUtils` | New method added. | - | `getGid(groupName: string): number` |
-
-Renamed and changed parameters. Added file owner and group options.
-
 
 
 ## [1.0.30] - 2024/6/27